requirements.txt
reflex>=0.4.7
openai>=1.14.0
reflex-chakra>=0.6.0
python-dotenv
ragflow-sdk==0.15.1